+++ modules/extension/src/decoders/iso-8859-1.ts
@@ -1,4 +1,4 @@
-import type { Decoder } from './type';
+import type { Decoder } from '@hex/types';
 
 const decoder: Decoder = (data) =>
 	[...data].map((byte) => {