+++ modules/extension/src/extension.ts
@@ -1,103 +1,13 @@
-import fs from 'fs';
-import path from 'path';
-
 import type { ExtensionContext } from 'vscode';
-import clearModule from 'clear-module';
 import { commands, window, workspace, StatusBarAlignment } from 'vscode';
 
-import type { DecoderItem } from './decoders';
 import { BinaryViewProvider } from './binary-view-provider';
+import { resolveCustomDecoders } from './custom-decoders';
 import { builtinDecoders, defaultDecoder } from './decoders';
-import { output } from './output';
 import { state } from './state';
 
-const customDecoderWatchers = new Set<fs.FSWatcher>();
-
-function resolveCustomDecoders() {
-	const customDecodersConfiguration = workspace
-		.getConfiguration('hexViewer')
-		.get<Record<string, string>>('customDecoders');
-
-	if (!customDecodersConfiguration) {
-		// eslint-disable-next-line @typescript-eslint/no-floating-promises
-		window.showErrorMessage(`Couldn't read custom decoders configuration.`);
-		return [];
-	}
-
-	const entries = Object.entries(customDecodersConfiguration);
-
-	if (!entries.length) {
-		return [];
-	}
-
-	const root = state.activeView
-		? workspace.getWorkspaceFolder(state.activeView.document.uri)
-		: workspace.workspaceFolders?.[0];
-
-	if (!root) {
-		// eslint-disable-next-line @typescript-eslint/no-floating-promises
-		window.showErrorMessage(`Couldn't resolve workspace root for custom decoders.`);
-		return [];
-	}
-
-	if (root.uri.scheme !== 'file') {
-		// eslint-disable-next-line @typescript-eslint/no-floating-promises
-		window.showWarningMessage(`Custom decoders are not supported in virtual workspaces.`);
-		return [];
-	}
-
-	return entries.reduce<DecoderItem[]>((result, [label, file]) => {
-		try {
-			const destinationPath = path.isAbsolute(file) ? file : path.join(root.uri.fsPath, file);
-
-			const currentWatcher = fs.watch(destinationPath, () => {
-				if (customDecoderWatchers.has(currentWatcher)) {
-					for (const watcher of customDecoderWatchers) {
-						watcher.close();
-					}
-
-					customDecoderWatchers.clear();
-
-					// eslint-disable-next-line @typescript-eslint/no-use-before-define
-					reloadDecoders();
-				}
-			});
-
-			customDecoderWatchers.add(currentWatcher);
-
-			clearModule(destinationPath);
-
-			const cwd = process.cwd();
-
-			process.chdir(path.isAbsolute(file) ? path.dirname(destinationPath) : root.uri.fsPath);
-
-			// eslint-disable-next-line @typescript-eslint/no-require-imports, @typescript-eslint/no-var-requires
-			const decoder: unknown = require(destinationPath);
-
-			process.chdir(cwd);
-
-			if (typeof decoder !== 'function') {
-				throw new TypeError(`Custom decoder '${label}' is not a function`);
-			}
-
-			result.push({
-				label,
-				decoder,
-			});
-		} catch (error) {
-			// eslint-disable-next-line @typescript-eslint/no-floating-promises
-			window.showErrorMessage(`Error resolving custom decoder '${label}'. See output for details.`);
-			// eslint-disable-next-line @typescript-eslint/restrict-template-expressions
-			output.appendLine(`${error}\n`);
-		}
-
-		return result;
-	}, []);
-}
-
 function reloadDecoders() {
-	state.decoderItems = [...builtinDecoders, ...resolveCustomDecoders()];
-	// const activeDecoder = state.decoderItems.find((item) => item.label === state.activeDecoderStatusItem.text);
+	state.decoderItems = [...builtinDecoders, ...resolveCustomDecoders(reloadDecoders)];
 	state.allViews.forEach((view) => {
 		const item = state.decoderItems.find(({ label }) => label === view.decoderItem.label);
 		view.decoderItem = item ?? defaultDecoder;
@@ -109,7 +19,7 @@ function reloadDecoders() {
 }
 
 export function activate(context: ExtensionContext): void {
-	state.decoderItems = [...builtinDecoders, ...resolveCustomDecoders()];
+	state.decoderItems = [...builtinDecoders, ...resolveCustomDecoders(reloadDecoders)];
 
 	state.activeDecoderStatusItem = window.createStatusBarItem(StatusBarAlignment.Right, 0);