+++ changelog.md
@@ -2,7 +2,7 @@
 
 All notable changes to the "hex-viewer" extension will be documented in this file.
 
-## [Unreleased]
+## [0.2.4] -- 2022-02-07
 ### Changed
 - Default decoder is ASCII now
 ### Added